LayoutTests/platform/ios-simulator/fast/canvas/canvas-skia-excessive-size-expected.txt

CONSOLE MESSAGE: line 20: Total canvas memory use exceeds the maximum limit (448 MB).
This test checks to see if the browser survives the attempted creation of an excessively large canvas.

Canvas 2d context = null!
Survived canvas creation attempt. Width = 134217728
